+++ src/tepkit/functions/vasp/__init__.py
@@ -1,7 +1,6 @@
 from ._system import stop_vasp, clear_outputs
 from .band_contour import band_contour
 from .get_band import get_band
-from .supercell import supercell
 from .get_neighbor_distances import get_neighbor_distances
 from .get_piezoelectric_stress_tensors import get_piezoelectric_stress_tensors
 from .get_thickness_info import get_thickness_info